requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14
django-cors-headers>=4.3
python-dotenv>=1.0

openai>=1.3,<2.0

whitenoise>=6.6
gunicorn>=21.2
